+++ functions/src/loyalty/index.ts
@@ -1,0 +1,6 @@
+export * from './types.js'
+export * from './customerRepository.js'
+export * from './loyaltyProgramRepository.js'
+export * from './loyaltyBalanceRepository.js'
+export * from './loyaltyTransactionRepository.js'
+export * from './saleLoyaltyResolver.js'